+++ src/lib/models/event-groups/create-event-group.test.ts
