+++ src/lib/models/event-groups/event-groups.d.ts
