+++ src/lib/models/event-groups/get-group-for-event.ts
