+++ src/lib/models/event-groups/get-group-id.ts
