+++ src/lib/models/event-groups/get-group-name.ts
