+++ src/lib/models/event-groups/get-last-event.ts
