+++ src/lib/models/event-groups/index.ts
